ENH: Refactor project structure and update dependencies; add main entry point
