+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
-current_version = 0.0.4
+current_version = 1.0.0
 commit = True
 tag = True
 
-[bumpversion:file:setup.py]
+[bumpversion:file:pyproject.toml]
