+++ karoloke/__init__.py
@@ -1,0 +1,4 @@
+from .jukebox import start_the_party
+
+if __name__ == "__main__":
+    start_the_party()
