+++ karoloke/general_config.py
@@ -4,4 +4,4 @@ SCORE_DB = './scores_videos'
 WND_NAME = 'Karol-Oke'
 MUSIC_BAR = './overlay/player_bar.png'
 ICONS_SET = './icons/'
-SCORE_BACKGROUND = './score'
+SCORE_BACKGROUND = './score_backgrounds'
