+++ pyproject.toml
@@ -16,6 +16,10 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.10"
+opencv-python = "^4.11.0.86"
+screeninfo = "^0.8.1"
+python-vlc = "^3.0.21203"
+pysimplegui = "^5.0.8"
 
 
 [tool.poetry.group.dev.dependencies]
@@ -25,6 +29,7 @@ pytest = "^8.3.4"
 pytest-cov = "^6.0.0"
 taskipy = "^1.14.1"
 ignr = "^2.2"
+bump2version = "^1.0.1"
 
 
 [tool.poetry.group.doc.dependencies]
