--- setup.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-
-from distutils.core import setup
-
-setup(
-    name='Karol-Oke',
-    version='0.0.4',
-    description='Simple framework for Karoke',
-    author='Antonio Filho',
-    author_email='acsenrafilho@gmail.com',
-    url='https://www.python.org/',
-)
